requirements.txt
aiohttp==3.8.1
colorama==0.4.5
discord==2.0.0
discord.py==2.0.1
python-dotenv==0.20.0